Initial monolithic fuzzball.ml version
Based on the oldest version in the Berkeley vine-execution SVN branch:
------------------------------------------------------------------------
r4130 | smcc | 2009-08-18 00:42:55 +0000 (Tue, 18 Aug 2009) | 5 lines

All the new code in the vine-execution project, as one giant file
which is the way I've been developing it so far. (Will reorganize in
the future). Also "trans_eval" [is not] intended as a permanent name or tool
interface.
------------------------------------------------------------------------
(In real history the source file was utils/trans_eval.ml, but this
artificial history moves it to the more modern location of
exec_utils/fuzzball.ml.)