Initial steps towards on-the-spot symbolic execution
------------------------------------------------------------------------
r4383 | smcc | 2009-10-21 00:34:26 +0000 (Wed, 21 Oct 2009) | 9 lines

Initial steps towards on-the-spot symbolic execution. Rewrite granular
memory to make the behavior when something is missing more flexible,
and add a string memory to match.
Make a datatype for register names, so that we don't have to pass
inconvenient Vine variables around everywhere.
Various small cleanups and reorganization of the main loop.
Add simple redirects for library functions we don't want to execute at
the moment.
------------------------------------------------------------------------