Make STP output easier to read and compare with stpvc
------------------------------------------------------------------------
r4946 | smcc | 2010-05-02 02:13:44 +0000 (Sun, 02 May 2010) | 8 lines

Fix a hard-to-find bug in the STP text output (arithmetic right shift
was failing to properly sign-extend because STP's >> doesn't shorten
in the way that its << lengthens). Along the way, add support for
running both stpvc and separate-process-STP to compare the results,
and various cleanups and extra simplifications to make the STP output
easier to read, such as using BOOLEANs for REG_1s and dropping
redundant variable number suffixes.
------------------------------------------------------------------------