Various changes in support of big/long-running programs
------------------------------------------------------------------------
r4985 | smcc | 2010-05-07 03:18:55 +0000 (Fri, 07 May 2010) | 7 lines

Add a more memory-efficient store_page for granular memory, and be
sure to use it on mmaps.
Add -env flag to set environment variables just for the subject program.
Add several new flags to make it easier to debug long-running programs.
Fix a bug that was leaving garbage in the high-order word of
st_blocks, and so causing glibc to suspect overflow.
------------------------------------------------------------------------