Add -symbolic-syscall-error and skipping options
------------------------------------------------------------------------
r5407 | smcc | 2010-11-10 02:39:15 +0000 (Wed, 10 Nov 2010) | 7 lines

Rename the poorly named -skip-call-addr option, and make a new option
to do more like what the old one sounded like, with regards to whether
the argument is the address of the call versus the address of the
called function.
Add an option to make a system call with symbolic arguments return a
given error code. (A different kind of workaround than concretization.)
------------------------------------------------------------------------