Improve condition flags and cjmp heuristic
------------------------------------------------------------------------
r5444 | smcc | 2010-11-17 07:36:03 +0000 (Wed, 17 Nov 2010) | 8 lines

Add an option -finish-on-nonfalse-cond, useful for terminating
bugfinding experiments.
Tweak how -extra-conditions are set, so they both mention input
variables and co-exist with -fulllen.
Call the cjmp_heuristic even if there's no choice, so it can keep
track of where we are if it wants. Also export an expression evaluator
for its use.
------------------------------------------------------------------------