Backport basic 64-bit host support
For ease of testing
------------------------------------------------------------------------
r5579 | smcc | 2011-01-12 01:58:09 +0000 (Wed, 12 Jan 2011) | 6 lines

Basic fixes to support running on a 64-bit Linux host:
Add a "-chroot" option to make it easy to get libraries (including a
SSE-free ld.so).
Fix some pattern matches on "-1" that fail when OCaml's "int" type is
bigger than 32-bits.
------------------------------------------------------------------------