Add "make bytecode" support
------------------------------------------------------------------------
r5816 | smcc | 2011-07-27 22:56:30 +0000 (Wed, 27 Jul 2011) | 5 lines

Build-process updates to switch to building just native code
executables with backtrace support by default. In support of this,
update the OCamlMakefile. If you still want bytecode, such as for
ocamldebug, it's now available under a "make bytecode" target.
------------------------------------------------------------------------
The main part of this change has been in the artificial history
from the start.