Basic Thumb instruction (16-bit ARM) support
------------------------------------------------------------------------
r5923 | smcc | 2011-09-27 23:43:51 +0000 (Tue, 27 Sep 2011) | 11 lines

Add basic support for Thumb (16-bit ARM) instructions.
One reason this is tricky is that VEX and some other parts of the
architecture use the convention of having the least significant bit of
the PC indicate whether you're in Thumb mode; matching this requires
some weird special cases in various places.
Another issue is that all Thumb instructions are potentially
conditional based on a dynamic condition; we generate a big decision
tree in each instruction to support this because we have to disable
VEX's optimization of this, but it's not yet tested because I still
need to get an example Thumb2 setup.
------------------------------------------------------------------------