Several useful new options, including -branch-preference
------------------------------------------------------------------------
r5941 | smcc | 2011-10-11 00:42:32 +0000 (Tue, 11 Oct 2011) | 9 lines

Give FuzzBALL the new options
  -skip-call-ret-region -skip-func-ret-region
  -concolic-cstring-file -branch-preference
which proved convenient in the OptiPNG example.
Make the register used in the -skip* options architecture based, so it
has a hope of working on ARM.
Invoke the cjmp heuristic when only one of the two labels in a cjmp is
an address, such as for x86 string instructions.
------------------------------------------------------------------------