Change the FPU emulator to run in user space
------------------------------------------------------------------------
r6039 | smcc | 2012-01-30 03:33:07 +0000 (Mon, 30 Jan 2012) | 3 lines

Make a version of the Linux FPU emulator that runs entirely in user
space (as an inlined interpreter).
------------------------------------------------------------------------
r6040 | smcc | 2012-01-30 03:49:44 +0000 (Mon, 30 Jan 2012) | 2 lines

Ignore the generated "tester" executable.
------------------------------------------------------------------------
r6065 | smcc | 2012-04-18 21:41:03 +0000 (Wed, 18 Apr 2012) | 3 lines

Add "-m32" so that this compiles on a 64-bit host as well (the code is
inherently 32-bit, including because it has lots of inlined assembly).
------------------------------------------------------------------------