Query-based bitwidth computation
------------------------------------------------------------------------
r6078 | smcc | 2012-04-28 03:06:04 +0000 (Sat, 28 Apr 2012) | 6 lines

Add a method to compute the size of a lookup table using decision
procedure queries, if we don't find a good bound syntactically. For
instance this lets us find a remaining table in the Poppler/Flate
example, though that appears not to help performance as much as I'd
hoped.
------------------------------------------------------------------------