Add -chroot support for connect and mkdir
------------------------------------------------------------------------
r6153 | smcc | 2012-09-19 20:28:37 +0000 (Wed, 19 Sep 2012) | 6 lines

Add chroot support for the paths in connect() to a UNIX-domain socket,
and in mkdir.
Note that the connect() change means you'll need to set up symlinks
in /tmp to run X programs inside a chroot, since X uses a Unix-domain
socket to connect to the server.
------------------------------------------------------------------------