Add -periodic-stats, with FM interface refactoring
------------------------------------------------------------------------
r6166 | smcc | 2012-10-18 21:37:44 +0000 (Thu, 18 Oct 2012) | 7 lines

Add the -periodic-stats option for seeing progress and getting things
like -gc-stats in the middle of long sequential runs.
This took more refactoring than I'd hoped, because the new
fm#add_extra_eip_hook method needs to take a function that has a
fragment machine as an argument, a kind of circularity that needs to
be handled in a particular way for OCaml's type system.
------------------------------------------------------------------------