Make region vars available and IVC more aggressive
------------------------------------------------------------------------
r6173 | smcc | 2012-11-13 23:26:39 +0000 (Tue, 13 Nov 2012) | 8 lines

Add region_vars to the input_dl, so that you can refer to, e.g.,
input0[1:reg32_t]:reg8_t in an -extra-condition.
Handle the case of memory variables in the evaluator (occurs when using
IVC and -symbolic-region)
Make simplify_with_callback call the callback more frequently: whenever
the expression is non-constant, not just when it's large. Has the effect
of making IVC more aggressive.
------------------------------------------------------------------------