Table stores and related improvements
------------------------------------------------------------------------
r6175 | smcc | 2012-11-13 23:36:31 +0000 (Tue, 13 Nov 2012) | 13 lines

Various improvements to the regions and tables code prompted by the
AbiWord XML case study:
Add store-to-table handling analogous to load-from-table handling,
causing each written location to become an if-then-else.
Improve the narrow_bitwidth calculation and term classification to fix some
bugs and better support new cases including XOR and if-then-else expressions
better. There's now an "AmbiguousExpr" classification for terms where we're
not sure whether it's a base or an offset, but it's a complex expression.
Add a maximum-value query analogous to the bitwidth one, for use to find
store bounds. Change the way both these queries work so they don't add to
the decision tree, but keep their own caches.
Refactor some code in support of the above.
------------------------------------------------------------------------