Add -stop-on-symbolic-syscall-args, -trace-guidance, bug fixes
------------------------------------------------------------------------
r6218 | smcc | 2013-02-27 16:48:23 +0000 (Wed, 27 Feb 2013) | 8 lines

Add a -stop-on-symbolic-syscall-args option to FuzzBALL, to recover a
previously-default behavior that is initially surprising, but works
well for programs that print error messages.
Fix a bug in -target-guidance that was causing the probability to be
always 100%, since I'd put the check in two places, and the random
check only at the wrong one.
Add a -trace-guidance option for debugging -target-guidance.
------------------------------------------------------------------------