Improved support for finishing cases
------------------------------------------------------------------------
r6219 | smcc | 2013-03-07 17:45:42 +0000 (Thu, 07 Mar 2013) | 6 lines

Fixes to allow reliably printing the assignment we're looking for at the
end of a successful search. When a checked condition "can be true or false",
take true on this path. And in that case as well as a final target match,
still finish up the path so -solve-final-pc applies and we can get more
statistics.
------------------------------------------------------------------------