Propagate DT heuristic more often
------------------------------------------------------------------------
r6222 | smcc | 2013-04-19 22:05:41 +0000 (Fri, 19 Apr 2013) | 6 lines

Propagate the heuristic every time it's updated, so we get the
right results even if it's not monotonic on a path.
Change to comparing only based on the maximum heuristic value: I
think this might help in some stuck cases, though this could use some
more analysis.
------------------------------------------------------------------------