Add input depth to guidance heuristic
------------------------------------------------------------------------
r6223 | smcc | 2013-04-19 22:07:38 +0000 (Fri, 19 Apr 2013) | 5 lines

Add a method to count how many input variables are mentioned in the
path condition, as a proxy for "how much of the input have we looked at"?
Set up a new utility formula based on the output/input ratio first and
the output match only secondarily. This helps some with the hex example.
------------------------------------------------------------------------