Add influence code that calls a query_engine directly
------------------------------------------------------------------------
r6225 | smcc | 2013-04-28 22:00:48 +0000 (Sun, 28 Apr 2013) | 7 lines

Start a new branch of the influence code that doesn't depend on the
old ValSet implementation; the old version now has "valset" in the
name.
In support of this, make some other code easier to reuse: conjoin and
disjoin don't need to be methods in formula_manager, and allow a
uniqueness suffix to be specified in construct_solver.
------------------------------------------------------------------------
Backported to replace the ValSet influence code, which is not
included. For simplicity of merging, this commit does not include
the cosmetic conjoin/disjoin change.