Update automake usage for future-proof-ness
Behavior change: add "subdir-objects", which puts the libasmir
.o files in the subdirectories of libasmir/src rather than all in
that directory. This is in some ways more logical, and is scheduled
to become the default behavior in the future.

Add the equivalent of "-Wall" for automake warnings, and add a
workaround for the one we currently trigger. The hope is this will
lead to catching issues sooner in the future.