Implement random(2cgc) and -symbolic-random
------------------------------------------------------------------------
r349 | smccamant | 2014-06-17 13:22:44 -0500 (Tue, 17 Jun 2014) | 7 lines

Basic implementation of random(2cgc) and -symbolic-random

Still untested: that will be the next step.
This touches a lot of files because I routed a new random_word method
through the DT and FM to ensure that the randomness is deterministic
when it's not symbolic.
------------------------------------------------------------------------