New option -finish-reasons-needed and related changes
Now that a -finish* condition doesn't immediately stop the path, it's
not a bug to run into more than one.
Another kind of generalization that can be useful is to require
multiple conditions before finishing exploration. The new -f-r-n option
gives a number of conditions: if that number are hit then the search
will really finish. (The old behavior corresponds to the default value
of 1.) I also considered negative conditions or requiring a
"conjunction" of conditions, which might also be useful, but this
seemed simplest.