Clear list of finishing reasons after each path
The previous change neglected to clear the list of reasons after each
path, so for instance you could hit a count of 2 by triggering the
same single condition on consecutive paths, which was not my intent.
Seeing that the state should be reset for each path reminded me
it was more logically part of the fragment machine object, so this
change also moves it there.