Add a -branch-preference-unchecked option
Like -branch-preference, but doesn't even check whether the
branch direction is feasible. This is rather unsafe, but it was
easy to implement. It would also be nice to have a version that
checked whether the requested direction was feasible, but did not
check its inverse: but I think that would require more complicated
decision tree changes.