Save EIPs in the decision tree
They're not used for anything explicitly at the moment, and their
presence arguably muddles the abstractions. But I think they will be
useful for tying the decision tree to the program for understanding,
and in particular for debugging situations where the decision tree
gets out of sync due to path determinism failures.