Add interval tree data structure
------------------------------------------------------------------------
r1301 | jthayer | 2014-11-10 08:43:58 -0600 (Mon, 10 Nov 2014) | 6 lines

Pulled the red-black tree into the interval tree definition, rather
than defining the interval tree as sitting on top of the red-black
tree.  Should give better performance, and we're not using red-black
trees anywhere else inside of fuzzball.
------------------------------------------------------------------------
r1303 | jthayer | 2014-11-10 10:14:38 -0600 (Mon, 10 Nov 2014) | 6 lines

Add some testing code to the interval_tree class.

Changed some of the range subtraction code, which the tests identified
as being buggy.
------------------------------------------------------------------------
r1304 | jthayer | 2014-11-10 11:47:18 -0600 (Mon, 10 Nov 2014) | 9 lines

Insertion code now properly (and eagerly) merges ranges in tree.

Although this means insertion is no longer log n time (it might be n
log n in the case where we add a range that consumes all other ranges
in the tree)  it does mean that at any point in time, we have the
minimal representation of the written ranges, and memory was the
limiting factor for the struct.
------------------------------------------------------------------------
r1306 | jthayer | 2014-11-10 12:24:45 -0600 (Mon, 10 Nov 2014) | 9 lines

Added remove range, fixed some issues / uncertainty wrt what it means
to remove [x,y] [a,b].  Specifically, it wasn't clear if we meant [a,
b], (a,b), [a,b) (a,b] before.  For this application, we want [a,b].
So the ambiguity was resolved that way.

Added an mli file so it's obvious what the interface is meant to be.
------------------------------------------------------------------------
r1309 | jthayer | 2014-11-10 13:51:58 -0600 (Mon, 10 Nov 2014) | 2 lines

Changed interval tree to work with int64s instead of ints.
------------------------------------------------------------------------