Fix interval trees for pointer management
------------------------------------------------------------------------
r1682 | jthayer | 2014-12-15 12:52:54 -0600 (Mon, 15 Dec 2014) | 8 lines

Fixes several issues in interval trees as they relate to their use in
pointer management.  The largest of these is that Maps are functional
datastructures, so I have to be careful to return the updated map so
it can be stored in pointer_management.ml.  Merges now also work on
abbutting intervals.  Safe_merge is much more careful about
tests as well.
------------------------------------------------------------------------