Change the rep. of 128-bit values to a pair of 64s
The expression type is named Vector, with the idea there might be other
sizes in the future. For now it should only exist transiently. With this,
we no longer need REG_128. My earlier plan had been to do this an
analogous transformation at the OCaml level, but after looking at it
more I decided it was easier to make the transformation earlier.