Add buffer location to random(2cgc) logging
------------------------------------------------------------------------
r1756 | jthayer | 2015-01-21 08:16:39 -0600 (Wed, 21 Jan 2015) | 10 lines

Changed fuzzball output for random calls, now it contains the location of the
buffer being read into as well as the number of bytes read into that
buffer.

This should let us catch simple challenge / response cases where, for
example, a session is identified by assigning it a random id that must
precede any message sent.  We just have to align the random bits with
outputs in the pov file.
------------------------------------------------------------------------