Track read/write ratio and other pointer management improvements
------------------------------------------------------------------------
r1757 | jthayer | 2015-01-21 08:42:51 -0600 (Wed, 21 Jan 2015) | 5 lines

Changed types on pointer_management errors (tuples to structs).
Shouldn't have impact on i/o of fuzzball, just wanted names to make it
easier to keep track of what data was where in warnings.
------------------------------------------------------------------------
r1758 | jthayer | 2015-01-21 09:58:38 -0600 (Wed, 21 Jan 2015) | 12 lines

fuzzball

When detecting a writing unallocated error, we now get a little more
information back from the interval_map.  In particular, we know if we
wrote before the start of an allocated region, after the end of an
allocated region, both of those things (on a single region), or if
we're writing to unallocated space.

This information isn't pushed back to the .info files yet, it's just
being staged for that at this point.
------------------------------------------------------------------------
r1759 | jthayer | 2015-01-21 10:23:56 -0600 (Wed, 21 Jan 2015) | 13 lines

When an illegal write happens, fuzzball now emits more information.

We generate a message containing the attempted write
range as well as the kind of illegal write (write to unallocated
memory, writes that intersect with but are not wholy contained by
allocated memory).  In the event that the write intersects allocated
memory, we emit the start and end address of one such segment (note,
not all such segments), as well as the direction of that intersection
(before, after, both).
------------------------------------------------------------------------
r1760 | jthayer | 2015-01-21 10:47:28 -0600 (Wed, 21 Jan 2015) | 1 line

adds find_all to the interval tree.  Important for being able to find all region
s that an illegal write manages to clobber
------------------------------------------------------------------------
r1761 | jthayer | 2015-01-21 11:23:40 -0600 (Wed, 21 Jan 2015) | 8 lines

Detect when a write exits an allocated area and enters another
allocated region (a write to A overflows and also touches all / part
of area B, C, etc..).

Outputs a json object to the info file that notes this has happened,
but doesn't yet contain the list of regions overwritten.
------------------------------------------------------------------------
r1762 | jthayer | 2015-01-21 13:23:42 -0600 (Wed, 21 Jan 2015) | 4 lines

List of overlapping segments comes back sorted (smallest index first)
------------------------------------------------------------------------
r1763 | jthayer | 2015-01-21 13:44:45 -0600 (Wed, 21 Jan 2015) | 7 lines

fuzzball

when a proposed write touches unallocated segments of memory and also
touches multiple allocated bits on the heap, we now generate a report
containing all of the ranges touched by the write.
------------------------------------------------------------------------
r1766 | jthayer | 2015-01-21 14:48:48 -0600 (Wed, 21 Jan 2015) | 7 lines

fuzzball

Push back information on the frequency of writes / reads from
interval tree back into pointer management.  Issues a warning to the
pov file if writes are more frequent than reads for a segment of memory.
------------------------------------------------------------------------