Add -read-write-ratio-warn option
------------------------------------------------------------------------
r1769 | jthayer | 2015-01-22 08:29:46 -0600 (Thu, 22 Jan 2015) | 9 lines

fuzzball

You can now set a threshold for the warning on read/write ratios for a
given piece of memory.  Previously the threshold was hard-coded to 2
(if you saw two or more writes before a read, a warning was issued).

The default has been pushed up from 2 to 5.
------------------------------------------------------------------------