Lopsized diamond detection for veritesting
------------------------------------------------------------------------
r1853 | jbenton | 2015-02-02 22:14:04 -0600 (Mon, 02 Feb 2015) | 19 lines

fuzzball

Lopsized diamond detection for veritesting.  Realized that regression
over the search tree was likely not the right idea because dangling
branches entail the existance of a second "exit point".

This simplifies matters, enforcing "only-one-branch" to avoid having
to worry about multiple exit points at the moment.  If we ever see
more than one branch, we will not declare it a diamond, however it may
be a "three-or-more-branch" diamond (e.g., switch
statements?). Expanding to that case shouldn't be particularly
difficult.  (The probably-very-usually-not-wrong assumption here is
that "dangling branches" will not simply halt.)

(Note: A bit of an interim commit -- the hash-hit match must kill its
child to avoid accidental encoding, which will make it resurrected
non-zombily.)
------------------------------------------------------------------------