Introduce temporary vars for float values in SMTLIB
This is needed because it's the only standard way to get a corresponding
bitvector. This requires changing the interface because we can't
introduce new variables in the middle of an expression. Take the
opportunity to remove the unhelpful (IMO) visitor design pattern,
introduce a pared-down interface, and remove unused helper functions.

(from branch f4651136 Sun, 8 Feb 2015 14:13:47 -0600)