Fix bug with null checks not starting symbolic execution
------------------------------------------------------------------------
r1983 | smccamant | 2015-02-22 23:07:50 -0600 (Sun, 22 Feb 2015) | 9 lines

Fix the case when a check for null is the first branch on symbolic
data on a path.

This was interacting badly with the optimization that tries to
set the res[e]t point at the first instruction with a branch, because
this check wasn't being recognized. Duplicate the setting of this
flag in more places in the hopes it won't be missed. (I also
renamed the interface slightly to make it IMO clearer.)
------------------------------------------------------------------------