Add basic protocol hints to FuzzBALL .info files
------------------------------------------------------------------------
r2219 | jthayer | 2015-03-30 16:37:17 -0500 (Mon, 30 Mar 2015) | 13 lines

Adds a couple of protocol hints to fuzzball's info output.

If we can detect that information being emitted is from outside the program, or
the result of a call to random,
we put additional information in the info file that contains:

What kind of hint we're giving (random emission or echo)
What memory range was emitted
Where the transmit happened in the program

Changes to cgc_syscalls provide the functionality, changes in pointer_management
 and interval_tree
fix some issues that were causing information to be lost on some writes.
------------------------------------------------------------------------