Fix save/restore of pointer_management state
------------------------------------------------------------------------
r2727 | smccamant | 2015-05-27 00:36:05 -0500 (Wed, 27 May 2015) | 6 lines

Save and restore the pointer management state by keeping copies internally
to the object, rather than by having the object have a deep copy method.
This is the way other FuzzBALL objects work, and avoids issues if the
object is aliased. The latter was a problem with the old code: many
places had references but only one reference was getting updated.
------------------------------------------------------------------------