Add basic implementation of SSE registers and FP ops
Primarily motivated by Clang's use of SSE FP, but can be the basis
for more SIMD and wide operations in the future. These changes
originally occurred on a branch along with other FP support, but
the history was a bit messy so this merges them as if they were
a fresh change.