More rules for simplifying shifts that end up as 0
Inspired by a situation in some x64 startup code seen on Ubuntu 14.04
that was reading from uninitialized memory. When the uninitialized
memory is made symbolic, these rules help FuzzBALL see that the symbolic
value is not actually used. Without this you couldn't set -fuzz-start-addr
without also -zero-memory as a workaround.