Remove base-address heuristic that caused DT inconsistency
This is the bug that inspired the debugging technique in my last commit.
Stop using the fact that a symbol has been used as the base address of
a synbolic region before to short-circuit the choice of which symbol
to use as a base in another access, since this information changes
across runs.