Rethink which args to mmap(2) should be int64
This should now follow our general rule that values that might be
pointers should be int64, while ones that will just be fairly small
integers (not even 32 bits on 32-bit) can be int. I think this should
be the best fix for the handling of FD -1 which was not working well
on some 64-bit systems.