Improvements in core dump and sysenter handling
Don't assert that all core notes are multiples of 4 bytes, since they
aren't.
Try a new approach to guess where a sysenter syscall will return to.
(Tends to apply now only when running from a core dump).
Add a -hwcap option to control AT_HWCAP, though it turns out that this
is not what controls what instruction is used for syscalls on x86.