Tweaks to behavior of satisfying assignment cache
Reduce the default size of the global cache: it tends to have less
benefit than the working cache, and checking a large global cache
can get expensive. A large global cache is probably still a win in
some cases, but a smaller size (100 entries) is a more balanced
default.
Collect basic hit-rate statistics.
Don't fill the cache if it's disabled anyway.
Don't add an entry to the working cache if it's identical to the
last-added entry.