Fix longstanding typo in "|" split term case
This code is currently commented out, perhaps because of a former
run-in with this bug that we never debugged. But it has a better chance
of working now if you want to try it.
Simplify in two places in the midst of symbolic address processing,
since with the fancier rules this can sometimes be helpful.
Add another set of simplification rules between casting and masking.