Solver support changes related to SMT arrays
Get a basic example working in both batch and incremental mode with
all our supported SMTLIB2 solvers.
Some more complicated refactoring was needed after I realized that
the definitions of t variables and tables can each refer to the
other, so we need to maintain the ordering among them as a generalization
of how we maintain the ordering among t variables.