Use .cvc instead of .stp as the extension for stp-external
The .stp extension was never special to STP AFAIK. Modern versions of STP
(since before 2.2.0) have changed the default format to SMTLIB2, so
they try to parse .stp file as if it were SMTLIB2 and fail. Modern
versions of STP also have a --CVC option, but relying on that would
break compatibility with older versions.