OCaml 4.04.1 build compatibility fixes
Fixes from one parser change and two new warnings, with the warnings
occuring in two places each. The parsing fix was a missing space that
was good to fix anyway, while the warning fixes are more subtle.

One warning asks us to not pattern match on the string arguments of
certain OCaml exceptions, because the OCaml developers want to be
able to change the strings. It's annoying that there's no replacement
way to get the same information, but in these cases, I think matching
all occurrences of the exception should be fairly safe.

The other warning is about potentially unintuitive behavior when a
single guard condition uses variables from multiple patterns. I think
this behavior would likely actually be fine for us. The recommended
fix would in our case involve a lot of cut-and-paste, so for now
I've just kept one of the four similar patterns. These special cases
are just optimizations anyway, and I think they're less used now that
we have a dedicated Ite expression type.