Short-circuit Ite evaluation with concrete condition
This is a slight optimization, but it is more important as a workaround
for cases where a memory load is not pure because of side-effects like
crashing if the address is null. For instance this can happen on ARM
when a pointer load is made conditional on a null check.

Cherry picked from part of the adaptor synthesis change 881a8fc3, but
with some whitespace and comment changes.