Translate guarded stores with branches not ITEs
This seems more correct because we don't want side effects of the store,
like memory access errors, when the guard is false.

This is an alternative fix for a problem that Vaibhav also encountered
in the work on adaptor synthesis 881a8fc3. But I think it's better not
to change the Vine IR semantics, but the translation instead.