Turn off numeric suffixes on variable printing
These have been a hassle for a while, but on mulling it over again
now I think they can probably go. They disambiguate so that two
variables that are internally distinct will always print differently,
but FuzzBALL is pretty good about giving unique string names to
variables so this is rarely a problem. For the solver interfaces
we use a hash table to exclude the suffixes only for the first variable
of a given name, which is a reasonable compromise but it would seem
wrong to have something like that in vine.ml. For the same reason it
wouldn't be convenient to have a command-line option.