Add a bitvector concatenate binop to the Vine IR
This operation was previously expressed with a combination of casts
and bitwise operators that was somewhat cumbersome. This addition
isn't completely orthogonal, but it's a common operation at the binary
level, which already exists in both VEX and SMTLIB, so it makes things
more direct to support it; it also makes formulas more readable IMO.
In text it's represented as infix "@", as in CVC/STP; for now only the
8 @ 8 = 16, 16 @ 16 = 32, and 32 @ 32 = 64 width combinations are
supported, because of Vine's limited bitvector sizes.

This commit has the addition to the Vine IR, additions of cases in all
the places the OCaml compiler catches, and a change for FuzzBALL to
use the new operator for the symbolic "assemble" operations. Still to
come are LibASMIR support, and simplification patterns.