Basic madvise(2), enhance mmap(2) flags and zero-fill
madvise is basically a no-op, though it specifically claims not to support
huge pages since they're probably not good for us.
Allow a number of more combinations of flags to mmap(2), leading to
the same basic behaviors of reading or zero-filling.
Use the store_page operation for zero-filling pages. This greatly reduces
memory usage compared to the old byte-by-byte filling behavior, especially
if the memory is unused.