Start adding specific x64 missing syscall messages
Putting these stubs in saves users doing a lookup in a kernel header file
when the unhandled syscall is encountered, and saves some time when
we get around to implementing a syscall. However with more than 300 of
these stubs or comments needed, I realized I don't have the time to
do them all now.