Implement -trace-{loads,stores} for Vinegrind
At the moment the loaded and stored values are printed differently
between Vinegrind and FuzzBALL, based on the difference between the
to_string functions of the concrete and symbolic domains. This makes
diff do a lot more extra work in the specific application I wanted this
for, comparing executions between Vinegrind and FuzzBALL, though it
still gets the right alignment in the end. I'm not sure if this should
be normalized in the future.