Add no-op implementation of prlimit64(2)
This matches our existing (get|set)rlimit implementations; prlimit64
is used instead by more recent libcs. I think it should work to use
the same implementation of either 32 or 64-bit, because the point of
prlimit64 is that it uses a 64-bit rlim_t even on 32-bit, though I
haven't actually tested with a 32-bit binary.