Refactor fxsave/fxrstor support to work with older VEX
Older versions had a single large dirty helper instead of smaller
ones, and the argument indexing was different.
After implementing this I discovered in testing that the version
compatibility benefit of this change is going to be limited: versions
of glibc new enough to use fxsave/fxrstor also use MPX, and the older
version (r2739) of VEX where I was testing the old fxsave/fxrstor
support also dies on MPX prefixes.