s/printf/eprintf/ on code that's been merged from the trunk
This class of differences in the CGC branch continue to be a hassle.
This change switches over a number of instances that hadn't been
changed before probably because they were in code that was added
with printf on the trunk, and then automatically merged. Note that
we haven't even bothered doing this with g_assert.