Add syscall stubs for mremap, fstatfs64, and 64-bit locking fcntls
These are all fake in one way or another. fstatfs64 gives hard-coded
results like the other statfs family calls. Setting locks is treated
as a no-op. And mremap always returns ENOSYS. I think in some cases
that's enough for users to fall back to doing nothing, or to just
doing a fresh mmap.