Add (currently unused) code to pad env or argv strings
This would be a low-tech way to reduce non-determinisim caused
by changes in the lengths of strings (for instance other environment
variables having different addresses). However it turned out not
to help in the case where I first thought it might, so leave it
disabled until we have a use case.