Adding a pair of meta options to FuzzBALL that control other options in an address range
This commit adds the `-turn-opt-on-range` and `-turn-opt-off-range`
options to FuzzBALL. They take arguments as `option:addr1:addr2`.
`-turn-opt-off-range` causes `option` to be turned on at all program
locations except those that are executed after `addr1` and before
`addr2`. Similarly `-turn-opt-on-range` turns `option` on only at
program locations that are reached after `addr2` and before `addr2`.