Two small compatibility fixes to configure.ac
For older systems, use "cd; git" instead of "git -C" when trying
to figure out the VEX version.

Add the "foreign" option to automake, to ask it not to complain about
non-GNU features of our build, such as certain documentation files.
At the moment I think the relevant files are in Git as empty, so
this would only be an issue if something else deletes them.